JRJC-116 - WorklogInput  and WorklogInputBuilder do not allow setting time spend in textual form (e.g. "1d 2h"). It requires specifying minutes, which is broken

+++ src/test/java/com/atlassian/jira/rest/client/internal/json/WorklogJsonParserTest.java
@@ -39,7 +39,7 @@ public class WorklogJsonParserTest {
         assertEquals(TestUtil.toDateTime("2010-08-17T16:35:47.466+0200"), worklog.getCreationDate());
         assertEquals(TestUtil.toDateTime("2010-08-17T16:35:47.466+0200"), worklog.getUpdateDate());
         assertEquals(TestUtil.toDateTime("2010-08-15T16:35:00.000+0200"), worklog.getStartDate());
-        assertEquals(60, worklog.getMinutesSpent());
+        assertEquals(60, worklog.getMinutesSpent().intValue());
         Assert.assertNull(worklog.getVisibility());
     }
 
@@ -56,7 +56,7 @@ public class WorklogJsonParserTest {
         assertEquals(TestUtil.toDateTime("2010-08-17T16:38:24.948+0200"), worklog.getUpdateDate());
         assertEquals(TestUtil.toDateTime("2010-08-17T16:37:00.000+0200"), worklog.getStartDate());
         assertEquals(Visibility.role("Developers"), worklog.getVisibility());
-        assertEquals(15, worklog.getMinutesSpent());
+        assertEquals(15, worklog.getMinutesSpent().intValue());
     }
 
     @Test
